feat: modularize project code
