+++ src/models/mod.rs
@@ -1,0 +1,6 @@
+// Авторские права (c) 2025 urdekcah. Все права защищены.
+//
+// Этот исходный код распространяется под лицензией AGPL-3.0,
+// текст которой находится в файле LICENSE в корневом каталоге данного проекта.
+pub mod api;
+pub mod weather;
